requirements.txt
click==6.7
coverage==4.4.1
fastcache==1.0.2
Jinja2==2.10
jsonschema==2.6.0
mmh3==2.3.1
mypy==0.550
pytest==3.3.1
pytest-cache==1.0
pytest-cov==2.5.1
pytest-flakes==2.0.0
pytest-mypy==0.3.0
pytest-pep8==1.0.6
pytest-sugar==0.9.0
python_arango==3.5.0
PyYAML==3.12
requests==2.18.4
Sphinx==1.6.5
sphinx-autobuild==0.7.1
sphinx-autodoc-typehints==1.2.3
sphinx-rtd-theme==0.2.4
sphinx-click==1.0.4
TatSu==4.2.2
twine==1.9.1
typing==3.6.2
ulid-py==0.0.6
